fix(build): add file Util.targets
In build process, the script must copy the dll files to 'tasks' folder.
The location of these files are stored in a group called 'AppAssemblies'
and this group is formed with .targets files presents in each project. So,
the Util.targets file must be created and added to build.proj file.