fix(build): adjustments to bump AssemblyInfo
The procedure to bump version of GeneralAssemblyInfo.cs file is done
after the compilation and dll generation of solution, and all the dlls are
built with wrong versions. So, this procedure (bump version) must be done
before solution build. In BumpVersion.Test project, must have a
'AfterBuild' task, with the procedures to: i) copy BumpVersion dlls to a
temp folder, ii) execute the tests, iii) bump version using assemblies
located on temp folder, to avoid process conflicts. With this, the
solution build procedure will use the correct GeneralAssemblyInfo.cs